Add comments

+++ app/Http/Controllers/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Comments;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CommentController extends Controller
 {
@@ -19,11 +20,15 @@ class CommentController extends Controller
     }
 
     public function delete(Comments $comments) {
-        if ($comments) {
-            if ($comments->delete()) {
-                return redirect()->back()->with('success', 'Комментарий удален');
-            }
-            return redirect()->back()->withErrors(['Failed' => "Не удалось удалить комментарий"]);
-        } 
+        if (Auth::user()->Role == 'Администратор' || Auth::id() == $comments->AuthorUserID) {
+            if ($comments) {
+                if ($comments->delete()) {
+                    return redirect()->back()->with('success', 'Комментарий удален');
+                }
+                return redirect()->back()->withErrors(['Failed' => "Не удалось удалить комментарий"]);
+            } 
+        } else {
+            return redirect()->back()->withErrors(['Failed' ,'Вы не администратор или комментарий вам не принадлежит']);
+        }
     }
 }
